services/analytic-service-python/requirements.txt
fastapi
uvicorn
motor
aiokafka
python-dotenv
pydantic
certifi
